gst/rtsp/: Make channel guint8 where possible.
Original commit message from CVS:
Patch by: Peter Kjellerstedt  <pkj at axis com>
* gst/rtsp/gstrtspsrc.h:
* gst/rtsp/rtspconnection.c: (rtsp_connection_receive):
* gst/rtsp/rtspmessage.c: (rtsp_message_init_data),
(rtsp_message_get_header):
* gst/rtsp/rtspmessage.h:
Make channel guint8 where possible.
Make rtsp_message_init_data() take the channel as a guint8.
* gst/rtsp/rtspdefs.c:
Fixed a typo: Timout -> Timeout
* gst/rtsp/rtspdefs.h:
Make RTSP_CHECK() behave as a statement.
* gst/rtsp/sdpmessage.c:
Avoid a compiler warning in INIT_ARRAY().
Fixes #437692.

+++ gst/rtsp/gstrtspsrc.h
@@ -90,7 +90,7 @@ struct _GstRTSPStream {
   gboolean      added;
 
   /* for interleaved mode */
-  gint          channel[2];
+  guint8        channel[2];
   GstCaps      *caps;
   GstPad       *channelpad[2];
